+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Requires PHP `8.4`
 - `Innmind\Server\Status\Server` is now a final class, all previous implementations are now flagged as internal
+- `Innmind\Server\Status\Server\Processes::all()` now returns an `Innmind\Immutable\Sequence`
+- `Innmind\Server\Status\Server\Disk::volumes()` now returns an `Innmind\Immutable\Sequence`
 
 ## 5.0.0 - 2025-05-08